back/user/service/application/requirements.txt

Flask==1.1.2
PyJWT==1.7.1
uWSGI==2.0.19.1
pymongo==3.11.0
passlib==1.7.2